Create a stand-alone `ColorExampleRow` component
that can be used in the `ColorRamps` as well as whatever new component handles the theme colors and background colors pages. It also doesn't have the same svg overflow and clipping hacks as the previous version (yay)
