+++ packages/docs/src/components/content/ColorRamps.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
-import classNames from 'classnames';
 import colorTokens from 'design-system-tokens/src/tokens/color';
-import { hexHasTransparency, pickTextColor } from 'design-system-tokens/src/lib/utility';
-import { HexValue } from 'design-system-tokens/src/lib/types';
+import ColorExampleRow from './ColorExampleRow';
 
 interface SwatchColor {
   name: string;
@@ -41,45 +39,14 @@ const ColorRamps = () => (
               <defs>{renderTransparencyPattern(patternId)}</defs>
               <rect x="0" y="0" width="100%" height="100%" fill={`url(#${patternId})`}></rect>
             </svg>
-            {swatchColors.map(({ name, value }) => {
-              const nameId = `color-name-${name}`;
-              const valueId = `color-value-${value}`;
-              const shortName = name.split('-')[1] ?? name;
-              const textColor = pickTextColor(
-                value as HexValue,
-                'var(--color-base-inverse)',
-                'var(--color-base)'
-              );
-              const codeStyle: React.CSSProperties = hexHasTransparency(value as HexValue)
-                ? {}
-                : {
-                    color: textColor,
-                    background: 'none',
-                  };
-
-              return (
-                <div className="c-color-ramp__item" key={`${name}-${value}`}>
-                  <svg aria-labelledby={`${nameId} ${valueId}`}>
-                    <rect x="0" y="-5%" width="100%" height="110%" fill={value} />
-                  </svg>
-                  <code
-                    className="ds-u-display--block"
-                    id={nameId}
-                    aria-describedby={valueId}
-                    style={codeStyle}
-                  >
-                    {shortName}
-                  </code>
-                  <code
-                    className="ds-u-display--block ds-u-margin-left--1"
-                    id={valueId}
-                    style={codeStyle}
-                  >
-                    {value}
-                  </code>
-                </div>
-              );
-            })}
+            {swatchColors.map(({ name, value }) => (
+              <ColorExampleRow
+                name={name}
+                value={value}
+                displayName={name.split('-')[1] ?? name}
+                key={`${name}-${value}`}
+              />
+            ))}
           </div>
         </div>
       );
