+++ packages/docs/src/styles/components/colorExampleRow.scss
@@ -1,0 +1,17 @@
+.c-color-example-row {
+  display: flex;
+  justify-content: space-between;
+  padding: $spacer-1;
+  position: relative;
+  text-align: right;
+
+  svg {
+    left: 0;
+    position: absolute;
+    top: 0;
+  }
+
+  code {
+    position: relative;
+  }
+}
