+++ packages/docs/src/styles/index.scss
@@ -11,6 +11,7 @@
 @import 'components/buttonMigrationTable.scss';
 @import 'components/buttonVariationsTable.scss';
 @import 'components/codeSnippet.scss';
+@import 'components/colorExampleRow.scss';
 @import 'components/colorSwatchList.scss';
 @import 'components/embeddedExamples.scss';
 @import 'components/exampleGrid.scss';
